refactor(account): move component to folder components
