+++ src/modules/account/Profile/components/ProfileForm/utils/mutationFetchPutProfileService/mutationFetchPutProfileService.ts
@@ -1,5 +1,5 @@
 import { fetchPutProfileService } from "modules/account/Profile/services";
-import { IProfileBody } from "modules/account/Profile/ProfileForm/types";
+import { IProfileBody } from "modules/account/Profile/components/ProfileForm/types";
 
 export const mutationFetchPutProfileService = ({ full_name }: IProfileBody) => {
 	return fetchPutProfileService({ body: { full_name } });
